requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
python-dotenv==1.0.1
mysql-connector-python==8.3.0
gunicorn==21.2.0
mysql-connector-python-rf==2.2.3
